Correct type specifications of parse/1,2
The type specification says the functions cannot err, though the code
clearly shows that as a possible outcome. This means users of the
dialyzer will get warnings in their code bases if they rely on the
error handling and matches on the error term.

Solve this by adding the right kind of type specification to the
different parts of the code base.